+++ FeedFeature/Feed API/FeedItemMapper.swift
@@ -23,10 +23,3 @@ internal class FeedItemsMapper {
         return root.items
     }
 }
-
-internal struct RemoteFeedItem: Decodable {
-    let id: UUID
-    let description: String?
-    let location: String?
-    let image: URL
-}
