+++ FeedFeature/Feed API/RemoteFeedItem.swift
@@ -1,0 +1,15 @@
+//
+//  RemoteFeedItem.swift
+//  FeedFeature
+//
+//  Created by Srinivasan Rajendran on 2020-12-25.
+//
+
+import Foundation
+
+internal struct RemoteFeedItem: Decodable {
+    let id: UUID
+    let description: String?
+    let location: String?
+    let image: URL
+}
